+++ .idea/workspace.xml
@@ -5,11 +5,21 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="81329931-dba5-4017-888a-9bb262c4d4a3" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/.idea/misc.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/misc.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/services/crud/crud_url/services/auth.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/crud/crud_url/services/auth.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/services/crud/crud_url/services/http_service.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/crud/crud_url/services/http_service.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/services/crud/crud_url/services/url_service.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/crud/crud_url/services/url_service.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/services/crud/requirements.txt" beforeDir="false" afterPath="$PROJECT_DIR$/services/crud/requirements.txt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/docker-compose.yml" beforeDir="false" afterPath="$PROJECT_DIR$/docker-compose.yml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/services/crud/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/crud/Dockerfile" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/services/crud/crud_url/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/crud/app/__init__.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/services/crud/crud_url/main.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/crud/app/main.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/services/crud/crud_url/schemas.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/crud/app/schemas.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/services/crud/crud_url/services/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/crud/app/services/__init__.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/services/crud/crud_url/services/auth.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/crud/app/services/auth.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/services/crud/crud_url/services/database.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/crud/app/services/database.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/services/crud/crud_url/services/http_service.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/crud/app/services/http_service.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/services/crud/crud_url/services/short_url_service.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/crud/app/services/short_url_service.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/services/crud/crud_url/services/url_service.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/crud/app/services/url_service.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/services/crud/run.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/crud/run.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/url_shortening_system.iml" beforeDir="false" afterPath="$PROJECT_DIR$/url_shortening_system.iml" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -37,22 +47,22 @@
     <option name="hideEmptyMiddlePackages" value="true" />
     <option name="showLibraryContents" value="true" />
   </component>
-  <component name="PropertiesComponent">{
-  &quot;keyToString&quot;: {
-    &quot;Docker.database: Compose Deployment.executor&quot;: &quot;Run&quot;,
-    &quot;Docker.docker-compose.yml.config1: Compose Deployment.executor&quot;: &quot;Run&quot;,
-    &quot;Python.run.executor&quot;: &quot;Run&quot;,
-    &quot;RunOnceActivity.OpenProjectViewOnStart&quot;: &quot;true&quot;,
-    &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
-    &quot;git-widget-placeholder&quot;: &quot;mvp&quot;,
-    &quot;kotlin-language-version-configured&quot;: &quot;true&quot;,
-    &quot;last_opened_file_path&quot;: &quot;/Users/lim/Documents/code/nanolink&quot;,
-    &quot;project.structure.last.edited&quot;: &quot;Modules&quot;,
-    &quot;project.structure.proportion&quot;: &quot;0.15&quot;,
-    &quot;project.structure.side.proportion&quot;: &quot;0.2&quot;,
-    &quot;settings.editor.selected.configurable&quot;: &quot;vcs.Subversion.SSH&quot;
+  <component name="PropertiesComponent"><![CDATA[{
+  "keyToString": {
+    "Docker.database: Compose Deployment.executor": "Run",
+    "Docker.docker-compose.yml.config1: Compose Deployment.executor": "Run",
+    "Python.run.executor": "Run",
+    "RunOnceActivity.OpenProjectViewOnStart": "true",
+    "RunOnceActivity.ShowReadmeOnStart": "true",
+    "git-widget-placeholder": "mvp",
+    "kotlin-language-version-configured": "true",
+    "last_opened_file_path": "/Users/lim/Documents/code/nanolink",
+    "project.structure.last.edited": "Modules",
+    "project.structure.proportion": "0.15",
+    "project.structure.side.proportion": "0.3402299",
+    "settings.editor.selected.configurable": "vcs.Subversion.SSH"
   }
-}</component>
+}]]></component>
   <component name="RecentsManager">
     <key name="MoveFile.RECENT_KEYS">
       <recent name="$PROJECT_DIR$/services/crud/crud_url/services" />
@@ -103,6 +113,7 @@
       <option name="PARENT_ENVS" value="true" />
       <envs>
         <env name="PYTHONUNBUFFERED" value="1" />
+        <env name="PYTHONPATH" value="services/crud/app" />
       </envs>
       <option name="SDK_HOME" value="" />
       <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/services/crud" />
@@ -157,6 +168,11 @@
       </deployment>
       <method v="2" />
     </configuration>
+    <list>
+      <item itemvalue="Docker.docker-compose.yml.config1: Compose Deployment" />
+      <item itemvalue="Docker.docker-compose.yml: Compose Deployment" />
+      <item itemvalue="Python.run" />
+    </list>
     <recent_temporary>
       <list>
         <item itemvalue="Python.run" />
