+++ services/crud/Dockerfile
@@ -1,0 +1,7 @@
+FROM python:3.12.2-slim
+WORKDIR /app
+COPY ./app /app
+COPY requirements.txt .
+RUN pip install --no-cache-dir -r requirements.txt
+EXPOSE 8000
+CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
