+++ services/crud/app/services/url_service.py
@@ -6,10 +6,10 @@ from motor.motor_asyncio import AsyncIOMotorClient
 from fastapi import Depends
 from bson import ObjectId
 
-from .database import get_database_service
-from .short_url_service import ShortUrlService, get_short_url_service
-from .http_service import HTTPService, get_http_service
-from ..schemas import UrlCreate, Url
+from services.database import get_database_service
+from services.short_url_service import ShortUrlService, get_short_url_service
+from services.http_service import HTTPService, get_http_service
+from schemas import UrlCreate, Url
 
 
 class UrlAlreadyExistsException(Exception):
