+++ services/crud/run.py
@@ -1,4 +1,14 @@
 import uvicorn
+import os
+import sys
+
+project_root = os.path.dirname(os.path.abspath(__file__))
+sys.path.insert(0, project_root)
 
 if __name__ == "__main__":
-    uvicorn.run("crud_url.main:app", host="0.0.0.0", port=8000, reload=True)
+    uvicorn.run(
+        "app.main:app",
+        host=os.getenv("CRUD_URL_HOST", "0.0.0.0"),
+        port=int(os.getenv("CRUD_URL_PORT", 3000)),
+        reload=True if os.getenv("ENV") == "dev" else False,
+    )
